Added ColorCMS function to Color field
ColorCMS returns HTML code with a visual representation and the HEX
code of the color for usage in CMS. More user-friendly as the editor
can actually see the color instead of guessing what the color code
means.

Can be used in `$summary_fields` of a DataObject:

```php
$summary_fields = [
    'Color.ColorCMS' => 'Color'
];
```

Result: GridFields will always render the visible color instead of only
the HEX code.